Added checks in installscript

+++ install.sh
@@ -12,6 +12,12 @@ if ! command -v go &> /dev/null; then
   exit 1
 fi
 
+# Check for kubectl installation
+if ! command -v kubectl &> /dev/null; then
+  echo "kubectl could not be found. Please install kubectl first."
+  exit 1
+fi
+
 # Initialize Go modules if not already initialized (in case of go.mod missing)
 if [ ! -f "go.mod" ]; then
   echo "No go.mod file found, initializing Go modules..."
@@ -30,12 +36,19 @@ go build $REPO_NAME.go
 echo "Installing the application globally..."
 #go install
 
-# Clean up: Delete the cloned repository
-echo "Cleaning up the repository..."
-mv $REPO_NAME ../$REPO_NAME.tmp
-cd ..
-rm -rf "$REPO_NAME"
-mv $REPO_NAME.tmp $REPO_NAME
+# Ask user if they want to delete the cloned repository
+read -p "Do you want to delete the cloned repository? (y/n): " delete_repo
+
+if [ "$delete_repo" = "y" ]; then
+  # Clean up: Delete the cloned repository
+  echo "Cleaning up the repository..."
+  mv $REPO_NAME ../$REPO_NAME.tmp
+  cd ..
+  rm -rf "$REPO_NAME"
+  mv $REPO_NAME.tmp $REPO_NAME
+else
+  echo "Keeping the cloned repository."
+fi
 
 # Output success message
 echo "Go application built and installed successfully!"
